se esta habilitando la extraccion de imagen de la galeria y camara

+++ lib/app/data/models/block_editor_model.dart
@@ -13,8 +13,8 @@ class BlockEditorModel {
 }
 
 class Details {
-  int height;
-  int width;
+  double height;
+  double width;
 
   Details({
     required this.height,
